requirements.txt
appdirs==1.4.4
asgiref==3.8.1
attrs==23.2.0
boto3==1.35.54
botocore==1.35.54
cffi==1.17.1
cmake==3.30.5
colorama==0.4.6
cryptography==43.0.3
defusedxml==0.8.0rc2
distlib==0.3.8
Django==4.2.3
django-allauth==65.0.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.15.4
fore==0.1
fsspec==2024.6.1
gunicorn==20.1.0
h11==0.14.0

idna==3.10
importlib-metadata==7.0.1
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
mysqlclient==2.2.4
networkx==3.3
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.1
parse==1.20.0
paypalrestsdk==1.13.3
pillow==11.0.0
platformdirs==4.2.2
pycparser==2.22
pyee==8.2.2
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyppeteer==1.0.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.3

selenium==4.17.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.5.1

tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.2

virtualenv==20.26.3
w3lib==2.1.2
websockets==10.4
wsproto==1.2.0
zipp==3.17.0

stripe==11.2.0